Fix handling of indefinite arrays/maps containing bools
Add a new helper to check if the next item is a Special::Break, and only
consume it if it is. Add new helpers to process arrays and maps using a
callback. Convert the deserializers for Vec and BTreeMap to use those.
Add tests for Vec<bool> and BTreeMap<bool>, which previously failed.